move out string function to new class
